add nvim and kitty to install

+++ install/prereq/aur.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 
-# install yay aur helper if not already available
-sudo pacman -Sy --needed --noconfirm git base-devel
-git clone https://aur.archlinux.org/yay.git
-cd yay
-makepkg -si --noconfirm
-cd -
-rm -rf yay
+if ! command -v yay1 &>/dev/null; then
+	echo "Installing yay aut helper."
+	# install yay aur helper if not already available
+	sudo pacman -Sy --needed --noconfirm git base-devel
+	git clone https://aur.archlinux.org/yay.git
+	cd yay
+	makepkg -si --noconfirm
+	cd -
+	rm -rf yay
+fi
